Refactor to use hiera data for defaults
This was an old Puppet 2.x modules that we had been carrying
forward; as a part of updating for recent operating system releases,
pull it up to at least Puppet 4.x.  (We can't bring it to 5.x yet
because many of our critical servers are still running 4.x with a 4.x
master.)

This splits out the old ntp::defaults manifest into the data and
updates pathnames and adjusts some code for FreeBSD 12.x where we are
planning a migration back to base ntpd rather than packaged ntpd.  We
add a FreeBSD-specific fact to indicate whether we are running on a
system with base ntpd installed, and new auto_leapseconds parameters
to work with startup scripts that automatically pull a leapseconds
file from an authoritative source.  (In the future, the version of the
leapseconds file in this module will be removed.)